feat: made the initial version of the firmware
The initial firmware isn't final by any means and is definitely subject
to change. There are still some things I want to achieve, those being:

- Swapping SUPER and CTRL when connecting to a Mac.
- Maybe making the timings a little different (slower).
- Adding a config layer where you can set a layout or default layer.
- Adding a Function (fn) layer.

We will see where the future will bring us. There is definitely one
problem with the firmware as of writing. For the firmware to work,
the "handness" needs to be written into the EPROM storage first after
which the firmware can be written deterministically. But that first
step does require you to take some indeterministic steps first.

For future versions of the keyboard I will make the "handness" be based
on a pin state (a jumper pulling it high or low). This will make the
firmware not require any EPROM steps first.