feat: added smarter defaults and better information output
Before if your system already had /dev/ttyACM0 then the script would fail
without the --serial-device parameter. This is now done better by looking for
the device smartly. It looks for the highest none existing serial device and
picks that one.

Also added more debug information if nix build fails, and would not tell you
why or that it failed. This is now fixed.