feat: added a 'remove other keyboards' flag to the install script
This flag removes other keyboards from the install repository (either locally or
globally). The point of this is that if you perform a fuzzy find in the
repository, the other repositories will not clog up your search results.

This flag is enabled by default for installing locally, and disabled by default
for installs that are performed globally.