feat(scripts): Added combined short flags support
So before if you provide the flags like this: `script -abc` instead of
`script -a -b -c` it would complain, saying it does not know the flag:
`-abc`. Now before parsing it split the short flags into their separated
versions. So now `script -abc` is equivalent to `script -a -b -c`.